+++ slices/mean_by.go
@@ -21,38 +21,22 @@ import (
 //	}
 //
 // Complexity: O(n)
-func MeanBy(slice interface{}, function func(interface{}) interface{}) interface{} {
+func MeanBy(slice interface{}, function func(interface{}) interface{}) float64 {
 	if !internal.SliceCheck(slice) {
 		panic("'slice' is not slice type")
 	}
 
 	sliceValue := reflect.ValueOf(slice)
-	sliceElementType := sliceValue.Type().Elem()
-
-	if sliceElementType.Kind() == reflect.Interface {
-		sliceElementType = reflect.TypeOf(1.0)
-	}
 
 	if sliceValue.Len() == 0 {
-		return reflect.Zero(sliceElementType).Interface()
+		return 0
 	}
 
 	floatType := reflect.TypeOf(1.0)
-	sum := reflect.Zero(floatType)
+	sum := 0.0
 	for i := 0; i < sliceValue.Len(); i++ {
-		element := reflect.ValueOf(function(sliceValue.Index(i).Interface()))
-
-		if internal.CanFloat(element.Interface()) {
-			sum = reflect.ValueOf(sum.Float() + element.Float())
-		} else {
-			sum = reflect.ValueOf(sum.Float() + element.Convert(floatType).Float())
-		}
+		sum += reflect.ValueOf(function(sliceValue.Index(i).Interface())).Convert(floatType).Float()
 	}
 
-	average := sum.Float() / float64(sliceValue.Len())
-	if average != float64(int(average)) && (sliceElementType.Kind() != reflect.Float32 || sliceElementType.Kind() != reflect.Float64) {
-		return average
-	}
-
-	return reflect.ValueOf(average).Convert(sliceElementType).Interface()
+	return sum / float64(sliceValue.Len())
 }