+++ slices/min_by.go
@@ -31,13 +31,16 @@ func MinBy(slice interface{}, function func(interface{}) interface{}) interface{
 		return nil
 	}
 
-	biggest := reflect.ValueOf(function(sliceValue.Index(0).Interface()))
+	biggest := reflect.ValueOf(function(sliceValue.Index(0).Interface())).Interface()
+	chosenElement := sliceValue.Index(0)
 	for i := 0; i < sliceValue.Len(); i++ {
-		element := reflect.ValueOf(function(sliceValue.Index(i).Interface()))
-		if res := internal.CompareNumbers(element.Interface(), biggest.Interface()); res == internal.Lower {
-			biggest = element
+		element := sliceValue.Index(i)
+		compareValue := reflect.ValueOf(function(element.Interface())).Interface()
+		if res := internal.CompareNumbers(compareValue, biggest); res == internal.Lower {
+			biggest = compareValue
+			chosenElement = element
 		}
 	}
 
-	return biggest.Interface()
+	return chosenElement.Interface()
 }