+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🐛 fix(slices-mean_by): MeanBy output expectation updated
 * 🐛 fix: bug fix on Mean and MaxBy and some changes on Intersection functions happened
 1.1.0 (2022-08-13)
 ------------------